refactor: move the business logic to the SourceCode entity
fix: remove all unused imports instead of just one

chore: update requirements
