+++ docs/requirements.txt
@@ -7,7 +7,7 @@
 babel==2.9.0              # via mkdocs-git-revision-date-localized-plugin
 beautifulsoup4==4.9.3     # via mkdocs-htmlproofer-plugin, mkdocstrings
 certifi==2020.12.5        # via requests
-chardet==3.0.4            # via requests
+chardet==4.0.0            # via requests
 click==7.1.2              # via mkdocs, nltk
 future==0.18.2            # via lunr
 gitdb==4.0.5              # via gitpython
@@ -16,7 +16,7 @@ htmlmin==0.1.12           # via mkdocs-minify-plugin
 idna==2.10                # via requests
 importlib-metadata==3.3.0  # via markdown
 jinja2==2.11.2            # via mkdocs
-joblib==0.17.0            # via nltk
+joblib==1.0.0             # via nltk
 jsmin==2.2.2              # via mkdocs-minify-plugin
 livereload==2.6.3         # via mkdocs
 lunr[languages]==0.5.8    # via mkdocs
@@ -38,7 +38,7 @@ pytkdocs==0.9.0           # via mkdocstrings
 pytz==2020.4              # via babel
 pyyaml==5.3.1             # via mkdocs
 regex==2020.11.13         # via nltk
-requests==2.25.0          # via mkdocs-htmlproofer-plugin
+requests==2.25.1          # via mkdocs-htmlproofer-plugin
 six==1.15.0               # via livereload, lunr
 smmap==3.0.4              # via gitdb
 soupsieve==2.1            # via beautifulsoup4
