+++ requirements-dev.txt
@@ -18,7 +18,7 @@ beautifulsoup4==4.9.3     # via mkdocs-htmlproofer-plugin, mkdocstrings
 black==20.8b1             # via -r requirements-dev.in
 certifi==2020.12.5        # via requests
 cfgv==3.2.0               # via pre-commit
-chardet==3.0.4            # via requests
+chardet==4.0.0            # via requests
 click==7.1.2              # via -r requirements.txt, autoimport, black, mkdocs, nltk, pip-tools, safety, yamlfix
 colorama==0.4.4           # via flakehell
 coverage==5.3             # via pytest-cov
@@ -61,7 +61,7 @@ importlib-metadata==3.3.0  # via flake8, flake8-comprehensions, flake8-simplify,
 iniconfig==1.1.1          # via pytest
 isort==5.6.4              # via -r requirements-dev.in, pylint
 jinja2==2.11.2            # via mkdocs
-joblib==0.17.0            # via nltk
+joblib==1.0.0             # via nltk
 jsmin==2.2.2              # via mkdocs-minify-plugin
 lazy-object-proxy==1.4.3  # via astroid
 livereload==2.6.3         # via mkdocs
@@ -99,13 +99,13 @@ pymdown-extensions==8.0.1  # via mkdocs-material, mkdocstrings
 pyparsing==2.4.7          # via packaging
 pytest-cov==2.10.1        # via -r requirements-dev.in
 pytest-pythonpath==0.7.3  # via -r requirements-dev.in
-pytest==6.2.0             # via -r requirements-dev.in, pytest-cov, pytest-pythonpath
+pytest==6.2.1             # via -r requirements-dev.in, pytest-cov, pytest-pythonpath
 pytkdocs==0.9.0           # via mkdocstrings
 pytz==2020.4              # via babel
 pyyaml==5.3.1             # via bandit, dparse, mkdocs, pre-commit, yamllint
 regex==2020.11.13         # via black, nltk
-requests==2.25.0          # via mkdocs-htmlproofer-plugin, safety
-ruyaml==0.19.2            # via yamlfix
+requests==2.25.1          # via mkdocs-htmlproofer-plugin, safety
+ruyaml==0.20.0            # via yamlfix
 safety==1.9.0             # via -r requirements-dev.in
 six==1.15.0               # via astroid, asttokens, bandit, flake8-debugger, livereload, lunr, pip-tools, virtualenv
 smmap==3.0.4              # via gitdb
@@ -125,5 +125,5 @@ yamllint==1.25.0          # via -r requirements-dev.in
 zipp==3.4.0               # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
-pip==20.3.1               # via pip-tools
+pip==20.3.3               # via pip-tools
 setuptools==51.0.0        # via flake8-annotations-complexity, flake8-expression-complexity, flake8-variables-names, ruyaml, safety, yamllint
