+++ tests/unit/test_services.py
@@ -4,7 +4,8 @@ from textwrap import dedent
 
 import pytest
 
-from autoimport.services import common_statements, fix_code
+from autoimport.model import common_statements
+from autoimport.services import fix_code
 
 
 def test_fix_code_adds_missing_import() -> None:
@@ -35,7 +36,7 @@ def test_fix_imports_packages_below_docstring() -> None:
     """Imports are located below the module docstrings."""
     source = dedent(
         '''\
-        """Module docstring
+        """Module docstring.
 
         """
         import pytest
@@ -43,7 +44,7 @@ def test_fix_imports_packages_below_docstring() -> None:
     )
     fixed_source = dedent(
         '''\
-        """Module docstring
+        """Module docstring.
 
         """
 
@@ -100,37 +101,32 @@ def test_fix_imports_type_hints() -> None:
     assert result == fixed_source
 
 
-def test_fix_substitutes_aliases() -> None:
-    """Missing import statements are loaded from the alias if there is a match."""
-    source = """getcwd()"""
-    aliases = {"getcwd": "from os import getcwd"}
-    fixed_source = dedent(
-        """\
-        from os import getcwd
-
-        getcwd()"""
-    )
-
-    result = fix_code(source, aliases)
-
-    assert result == fixed_source
-
-
-def test_fix_doesnt_fail_if_object_not_in_aliases() -> None:
-    """If no aliases are found for the required object, do nothing."""
-    source = """getcwd()"""
-    aliases = {"Dict": "from Typing import Dict"}
-
-    result = fix_code(source, aliases)
-
-    assert result == source
-
-
 def test_fix_removes_unneeded_imports() -> None:
     """If there is an import statement of an unused package it should be removed."""
     source = dedent(
         """\
         import requests
+        foo = 1"""
+    )
+    fixed_source = "foo = 1"
+
+    result = fix_code(source)
+
+    assert result == fixed_source
+
+
+def test_fix_removes_multiple_unneeded_imports() -> None:
+    """
+    Given: A source code with multiple unused import statements.
+    When: fix_code is run.
+    Then: The unused import statements are deleted.
+    """
+    source = dedent(
+        """\
+        import requests
+        from textwrap import dedent
+
+        from yaml import YAMLError
         foo = 1"""
     )
     fixed_source = "foo = 1"
